de-hardcodes job icons (#64193)
I made this because I was just messing with jobs and remembered about jobs' tgui prefs being de-hardcoded so thought about doing it for the in-game jobs as well.
Why It's Good For The Game

Not good for the game, but is just a QoL for downstreams who have custom jobs.
Changelog

Not necessary.