Update the Play Internet Sound Verb (#72805)
This changes it to expose additional metadata as well as remove unused
Metadata that wasn't provided by youtube-dl, It also changes the widget
itself to display a little extra info on the currently playing track.

Also adds in a new prompt to either show or hide the name of the admin
who played the music

It will now show additional info to players when a admin plays a song
using the verb, such as making the link always available to the player
so if they like the song they dont need to ask for the link, as well as
other info that might be scraped by youtube-dl so players know if you
simply queued a 2 minute long song or a 10 minute long one as well other
simple info such as artist and album if that info is available.

And now there is the option to show players who played songs, or to keep
it anonymous

<details>
<summary>Screenshots</summary>

![image](https://user-images.githubusercontent.com/2568378/213298504-1aa5bf74-9606-4214-a908-82cf275850c5.png)

![image](https://user-images.githubusercontent.com/2568378/213350021-5eaf1510-ccd7-45c1-9b84-1210023e36bf.png)

![image](https://user-images.githubusercontent.com/2568378/213298581-bbee80f7-6c1f-491c-b881-a954ccc61a80.png)

![image](https://user-images.githubusercontent.com/2568378/213346931-33884c65-fadc-46f8-ba8d-1f66722e0479.png)

![image](https://user-images.githubusercontent.com/2568378/213346943-b61cd243-ae81-4dab-b8d1-62642c85c4e5.png)

</details>

:cl:
admin: Modified the Play Internet Sound Verb so it now displays extra
info about any song being actively played as well as showing the link
/:cl:

Co-authored-by: san7890 <the@san7890.com>
Co-authored-by: Fikou <23585223+Fikou@users.noreply.github.com>