Request Internet Sound Optional Credit User  (#76453)
## About The Pull Request

What this PR does is it adds a small additional feature to the Request
Internet Sound verb to give users who request music an option to credit
themselves for the song request. It will use character name of the
person who submits the request, with the option to be anonymous. The
default will make users anonymous.

## Why It's Good For The Game

This differentiates songs that where requested by users and songs that
admins themselves played. And allows users to give themselves credit for
their 'Good' (or bad) music preferences.

<details>
<summary>Screenshot</summary>


![dreamseeker_OQyx1sZ689](https://github.com/tgstation/tgstation/assets/2568378/806347a7-a930-4993-b2d0-b1890021c662)

</details>

## Changelog

:cl:
qol: Request Internet Sound now has the option to credit the person who
requested the Sound. Defaults to anonymous.
/:cl: