Stop stack tracing if ezdb daemon is already running (#76555)
This is an annoyance for me because I often kill the game but not the
EZDB daemon (leaving the cmd window open), and it just triggers this
stack trace while I have a debugger running. It's still able to connect
to the daemon in-game and functions perfectly fine, just hate dealing
with a completely needless stack_trace() every single time.

![image](https://github.com/tgstation/tgstation/assets/34697715/f8ae3cec-1cf6-4e70-b9bc-c7088121b85c)

![image](https://github.com/tgstation/tgstation/assets/34697715/a13774b8-addf-43fd-82f0-3d0e644d2dc1)

LET ME WORK!!! PLEASE!!!
Nothing that players should care about.