Standardizes a majority of the GAGS configs for Clothing (#76384)
Yikes this stuff is messy to work with...

Fairly comprehensive changelist:
- Removed a LOT of redundant .json files, huzzah!
- Moves the Large Scarf into the normal Scarves .json since they're both
2 component states and can thus share a file
- Renamed almost every single config to have the same naming schemes
("Name (Type)")
- Subtyped a majority of the configs to related ones, inheriting where
possible
- Renamed some icons to account for now-inherited .jsons where needed

---

**A relevant question for coders/maintainers:**
As of now, the .json configs are structured with a few rules in place:
- a .json config's states MUST have the same amount of colored component
states
- a .json config can not have states that are not buildable via a linked
.dmi

The second one is proving troublesome; namely, for toggleable clothing.
The object sprites of a majority of these don't change, but the worn
sprites have _d added to the end. For a majority of the cases I've been
able to just share a .json for the object and worn GAGS configs, but for
every toggleable item (such as jumpsuits) they need to keep their own
separate .json just to add a new state.
I'm not knowledged in coding proper systems like this, so I was just
wanting to ask, is there a way that GAGS's .json configs can have
toggle-states without assembling one for the objects?
Code standardization and legibility yahoo
:cl:
refactor: refactored some code relating to greyscale clothing - please
report relevant issues on github!
/:cl: