Fixes music importing newlining (#82152)
Makes the instrument song importing tgui input text use multiline which
means BPM is now on its own line instead of taking the first line of the
song with it, getting it cut from the import.

Closes https://github.com/tgstation/tgstation/issues/82148

:cl:
fix: Importing songs into instruments no longer cuts the first line of
the song.
/:cl: