Fixes Midi2Piano Tool (#85967)
I put back pclip turns out the other way I was using copy to clipboard
did not work well, Made the program run once and it will always be on
top so you do not lose the window, and it will tell you when it's copied
the sheet music to your clipboard.

Fixes The Midi2Piano Tool also now you can't lose it under other
windows.