Mentor fixes and TGUIification (#12255)
* Update mrat.dm (fixes wrong user alert to remove mrat)

changes a "friend" to "friend.owner" in the alert.

* TGUIfied some of the mentor verbs

changed some inputs related to mentor actions from the old style to TGUI (the prompt or both mentor and mentoree to leave/kick the mrat, the name requrest for mrat on the mentor side, and the actual mhelp which seemed to be bugged).

* Makes msay chat hotpink again

Just as the title says

* Undoes the TGUI fix for the mentorhelp verb

It's too big in scope for this PR.