Object of obessession + alluring
Vampires with this quirk only get the full benefit of feeding from their chosen target.
Only humans with the alluring quirk can be chosen.
If there are no alluring humans, it will assign the first one to join.
You can feed off other targets, but only for a quarter of the benefit.
Also makes it so potent blood heals vampires.

If your obsession dies or leaves a new one is picked.

The primary intention behind this is to pull humanity out of the role of mere spectators and into the beating heart of the supernatural narrative.