Expanding roles for lesser played clans.
I personally believe the size of a clan shouldn't affect the roles that one has access to, opening up possibilities for roleplay opportunities should encourage more creative and overall fun interactions. One issue I've seen raised is that most of the player base are unaligned with the Anarch's, Camarilla, or otherwise and expanding on the options for lesser clans could expand the playerbase for the clan in question but also have more individuals for otherwise necessary roles.