SpeedBike to Motorcycle
This changes the first speedbike in bike.dmi from a hover bike to a street cruiser bike.

Lowers the speed by using .5 move delay.
Changes the trail to a smoke trail.

Fixes an issue where the icons the bike would select where not in the right order. By putting Bike_background icons on the obj layer, below the player, and bike parts that overlay will clip over the character.